BLE: Split parsing into new class
